Consolidate startEmbeddedTomcat and stopEmbeddedTomcat into StartTomcat and StopTomcat, respectively and switch logic there based on the presence of the useEmbeddedTomcat property.
